Move range.go into sharding package to avoid import cycles
